Move utils to Suchit's suggested location, fix comments
