+++ cypress/integration/core-opensearch-dashboards/opensearch-dashboards/apps/query_enhancements/saved_search.spec.js
@@ -19,7 +19,7 @@ import {
   verifyDiscoverPageState,
   verifySavedSearchInAssetsPage,
   postRequestSaveSearch,
-} from './utils/saved_search';
+} from '../../../../../utils/apps/query_enhancements/saved_search';
 
 export const runSavedSearchTests = () => {
   describe('saved search', () => {
